requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
coverage==7.6.10
crispy-bootstrap5==2023.10
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==5.1.5
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-filter==24.3
django-haystack==3.2.1
django-location-field==2.7.2
django-otp==1.3.0
django-watson==1.6.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
firebase-admin==6.5.0
flake8==5.0.4
flake8-docstrings==1.7.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
jiter==0.8.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
openai==1.59.6
packaging==24.2
pep8-naming==0.13.3
phonenumbers==8.13.53
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==7.1.3
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
qrcode==8.0
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
snowballstemmer==2.2.0
sqlparse==0.5.3
telegram==0.0.1
tomli==2.2.1
tqdm==4.67.1
twilio==9.4.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
wget==3.2
yarl==1.18.3
zope.interface==7.2
